+++ glycowork/glycan_data/loader.py
@@ -1,3 +1,5 @@
+import os
+os.environ["HF_HUB_DISABLE_SYMLINKS_WARNING"] = "1"
 import re
 import json
 import pickle